Move base analytic unit model class to models/analytic_units
