+++ server/src/models/threshold_model.ts
@@ -1,4 +1,4 @@
-import { AnalyticUnitId } from './analytic_unit_model';
+import { AnalyticUnitId } from './analytic_units/analytic_unit_model';
 
 import { Collection, makeDBQ } from '../services/data_service';
