+++ server/src/routes/threshold_router.ts
@@ -1,6 +1,6 @@
 import * as AnalyticsController from '../controllers/analytics_controller';
 
-import { AnalyticUnitId } from '../models/analytic_unit_model';
+import { AnalyticUnitId } from '../models/analytic_units/analytic_unit_model';
 import * as Threshold from '../models/threshold_model';
 
 import * as Router from 'koa-router';
